Refactor:PlaceDetail 관련 폴더 구조 수정
