+++ app/feature/place/components/PlaceDetail/organisms/PlaceKeywordSummaryLine.tsx
