+++ app/feature/place/components/PlaceDetail/organisms/PlaceKeywordSummaryPurpose.tsx
