+++ app/feature/place/components/PlaceDetail/organisms/PlacePurposeBox.tsx
